수정, 삭제권한 체크

+++ src/main/java/org/koreait/controller/ArticleController.java
@@ -18,11 +18,24 @@ public class ArticleController {
         this.articleService = new ArticleService(con);
     }
 
-    public void doAction() {
-
+    public void doAction(String actionMethod, int idx) {
+        if (MemberController.getLoginedMember() == null){
+            System.out.println("로그인 후에 이용해주세요.");
+            return;
+        }
+        switch (actionMethod) {
+            case "write" -> doWrite();
+            case "list" -> doList();
+            case "delete" -> doDelete(idx);
+            case "update" -> doUpdate(idx);
+            case "detail" -> viewDetail(idx);
+            default -> System.out.println("올바른 명령어를 입력해주세요.");
+        }
     }
 
     public void doWrite() {
+        Map<String, Object> nowMember = MemberController.getLoginedMember();
+        int nowMemberId = Integer.parseInt(nowMember.get("id").toString());
         System.out.println("게시글을 작성합니다.");
         System.out.print("title : ");
         String title = Container.getSc().nextLine();
@@ -36,32 +49,44 @@ public class ArticleController {
             System.out.println("내용은 필수 입력 사항입니다.");
             return;
         }
-        int lastId = articleService.insertArticle(title, body);
+        int lastId = articleService.insertArticle(title, body, nowMemberId);
         System.out.printf("%d번 게시글이 작성되었습니다.\n", lastId);
     }
 
-    public void doList() throws SQLException {
-        System.out.println("  번호  /      일시      /      제목      /     내용     ");
+    public void doList() {
+        System.out.println("  번호  /      일시      /   작성자   /      제목      /     내용     ");
         List<Map<String, Object>> rs = articleService.viewArticleList();
         for (Map<String, Object> map : rs) {
             String displayId = map.get("id").toString();
             String displayRegDate = dateTimeForDisplay(map.get("regDate").toString());
+            String displayNickName = subForDisplay(map.get("nickName").toString());
             String displayTitle = subForDisplay(map.get("title").toString());
             String displayBody = subForDisplay(map.get("body").toString());
 
             if (displayId.length() == 1) displayId = "0" + displayId;
 
-            System.out.printf("   %s   /   %s   /     %s     /     %s     \n",
-                    displayId, displayRegDate, displayBody, displayTitle);
+            System.out.printf("   %s   /   %s   /   %s   /     %s     /     %s     \n",
+                    displayId, displayRegDate, displayNickName, displayBody, displayTitle);
         }
     }
 
     public void doDelete(int idx) {
+        Map<String, Object> nowMember = MemberController.getLoginedMember();
+        String nowMemberId = nowMember.get("id").toString();
+
         if (idx == 0) System.out.println("올바른 id를 입력해주세요.");
         else {
-            int row = articleService.deleteArticle(idx);
-            if (row == 0) System.out.printf("%d번 article은 없습니다.\n", idx);
-            else System.out.printf("%d번 article이 삭제되었습니다.\n", idx);
+            Map<String, Object> checking = articleService.viewOneArticle(idx);
+            if (checking.isEmpty()) {
+                System.out.printf("%d번 article은 없습니다.\n", idx);
+                return;
+            }
+            if (!checking.get("author").equals(nowMemberId)) {
+                System.out.println("자신이 작성한 글만 지울 수 있습니다.");
+                return;
+            }
+            articleService.deleteArticle(idx);
+            System.out.printf("%d번 article이 삭제되었습니다.\n", idx);
         }
     }
 
@@ -72,6 +97,7 @@ public class ArticleController {
             if (rs.get("id") == null) System.out.printf("%d번 article은 없습니다.\n", idx);
             else {
                 System.out.println("번호 : " + rs.get("id"));
+                System.out.println("작성자 : " + rs.get("nickName"));
                 System.out.println("작성 일시 : " + rs.get("regDate"));
                 System.out.println("최종 수정 날짜 : " + rs.get("updateDate"));
                 System.out.println("제목 : " + rs.get("title"));
@@ -81,12 +107,20 @@ public class ArticleController {
     }
 
     public void doUpdate(int idx) {
+        Map<String, Object> nowMember = MemberController.getLoginedMember();
+        String nowMemberId = nowMember.get("id").toString();
+
+
         if (idx == 0) System.out.println("올바른 id를 입력해주세요.");
         else {
-            Map<String, Object> rs = articleService.viewOneArticle(idx);
-            if (rs.get("id") != null) {
-                String oldTitle = (String) rs.get("title");
-                String oldBody = (String) rs.get("body");
+            Map<String, Object> checking = articleService.viewOneArticle(idx);
+            if (checking.get("id") != null) {
+                if (!checking.get("author").equals(nowMemberId)) {
+                    System.out.println("자신이 작성한 글만 수정할 수 있습니다.");
+                    return;
+                }
+                String oldTitle = (String) checking.get("title");
+                String oldBody = (String) checking.get("body");
                 System.out.println("기존 title : " + oldTitle);
                 System.out.println("기존 body : " + oldBody);
                 System.out.print("새로운 title : ");
